# swagger-raw.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Novvi API",
    "description": "API REST do domínio Novvi (usuários, funcionários, produtos e pedidos) - FIAP Challenge, Clean Architecture.",
    "version": "v1"
  },
  "paths": {
    "/api/Funcionarios": {
      "get": {
        "tags": [
          "Funcionarios"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuncionarioResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuncionarioResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuncionarioResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Funcionarios"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FuncionarioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FuncionarioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FuncionarioRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Funcionarios/{id}": {
      "get": {
        "tags": [
          "Funcionarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Funcionarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FuncionarioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FuncionarioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FuncionarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Funcionarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pedidos": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PedidoResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PedidoResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PedidoResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pedidos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pedidos/{id}/pagamento": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagamentoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PagamentoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PagamentoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FuncionarioRequest": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "nullable": true
          },
          "salario": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "FuncionarioResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "salario": {
            "type": "number",
            "format": "double"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PagamentoRequest": {
        "type": "object",
        "properties": {
          "tipoPagamento": {
            "$ref": "#/components/schemas/TipoPagamento"
          }
        },
        "additionalProperties": false
      },
      "PagamentoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tipoPagamento": {
            "$ref": "#/components/schemas/TipoPagamento"
          },
          "horario": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PedidoRequest": {
        "type": "object",
        "properties": {
          "idUsuario": {
            "type": "string",
            "format": "uuid"
          },
          "idFuncionario": {
            "type": "string",
            "format": "uuid"
          },
          "produtosIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PedidoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "frete": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "idUsuario": {
            "type": "string",
            "format": "uuid"
          },
          "idFuncionario": {
            "type": "string",
            "format": "uuid"
          },
          "produtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProdutoResponse"
            },
            "nullable": true
          },
          "pagamento": {
            "$ref": "#/components/schemas/PagamentoResponse"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProdutoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "preco": {
            "type": "number",
            "format": "double"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "TipoPagamento": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  },
...
}